refactor(desktop): extract the release cleanup lifecycle into updater/release-lifecycle
Mechanical move out of updater.ts: the release cleanup.json descriptor
types and guards, deprecate/remove scanning with deferred-failure retry,
the mkdir-based lifecycle lock, and the .back directory sweep move to
updater/release-lifecycle.ts. The DesktopUpdater interface type stays in
updater.ts; orphaned imports pruned. This completes the split: updater.ts
is now the assembly layer plus the state machine and download/install
orchestration.
